+++ system/src/Grav/Common/Themes.php
@@ -311,17 +311,23 @@ class Themes extends Iterator
      */
     protected function autoloadTheme($class)
     {
-        /** @var UniformResourceLocator $locator */
-        $locator = $this->grav['locator'];
-
         $prefix = "Grav\\Theme";
         if (false !== strpos($class, $prefix)) {
             // Remove prefix from class
             $class = substr($class, strlen($prefix));
 
-            // Replace namespace tokens to directory separators
+            // Try Old style theme classes
             $path = strtolower(ltrim(preg_replace('#\\\|_(?!.+\\\)#', '/', $class), '/'));
-            $file = $locator->findResource("themes://{$path}/{$path}.php");
+            $file = $this->grav['locator']->findResource("themes://{$path}/{$path}.php");
+
+            // Load class
+            if (file_exists($file)) {
+                return include_once($file);
+            }
+
+            // Replace namespace tokens to directory separators
+            $path = $this->grav['inflector']->hyphenize(ltrim($class,"\\"));
+            $file = $this->grav['locator']->findResource("themes://{$path}/{$path}.php");
 
             // Load class
             if (file_exists($file)) {